+++ packages/elements/mention/src/getMentionOnSelectItem.ts
@@ -1,7 +1,9 @@
 import {
-  ComboboxItemData,
   ComboboxOnSelectItem,
   comboboxStore,
+  Data,
+  NoData,
+  TComboboxItem,
 } from '@udecode/plate-combobox';
 import { getBlockAbove, insertNodes } from '@udecode/plate-common';
 import {
@@ -14,18 +16,18 @@ import { ELEMENT_MENTION, ELEMENT_MENTION_INPUT } from './defaults';
 // FIXME: Cannot figure out the TS for this to work with insertNodes
 // import { MentionNodeData } from './types';
 
-export interface CreateMentionNode {
-  (item: ComboboxItemData): Record<string, unknown>;
+export interface CreateMentionNode<TData extends Data> {
+  (item: TComboboxItem<TData>): Record<string, unknown>;
 }
 
-export const getMentionOnSelectItem = ({
+export const getMentionOnSelectItem = <TData extends Data = NoData>({
   pluginKey = ELEMENT_MENTION,
   createMentionNode = (item) => ({ value: item.text }),
   insertSpaceAfterMention,
 }: {
-  createMentionNode?: CreateMentionNode;
+  createMentionNode?: CreateMentionNode<TData>;
   insertSpaceAfterMention?: boolean;
-} & PlatePluginKey = {}): ComboboxOnSelectItem => (editor, item) => {
+} & PlatePluginKey = {}): ComboboxOnSelectItem<TData> => (editor, item) => {
   const targetRange = comboboxStore.get.targetRange();
   if (!targetRange) return;