+++ packages/marks/font-ui/src/ColorPicker/ColorType.ts
@@ -1,0 +1,5 @@
+export type ColorType = {
+  name: string;
+  value: string;
+  isBrightColor: boolean;
+};